This commit includes a change in the names of the scripts, to have an external commit to execute the functions.
BUT it is not working right now form spyder (only from the terminal) (in the case of the execute).
So, right now, I am only running the functions_toyModel.py script. It has a if name = main to test the funtions.

I also include a first general dynamic function w/o the contact and harvest module defined.