En este commit solo arreglé cosas del T --> time y volví a separar el ejecutor de las fucniones. Esto es porque quiero poder hacer lo del realtive path que solo parece funcionar con eso.
Relative path runs well from terminal, but not from the code per se. (this has to be related with the virtual env)

Now, I0ll create a branch were I will try to modify the tmeporal DF instead of doing so many functions in the contact module.
I'm not doing the harvest module yet.