WUJUUU
This commit has the new general HM_general in the functions.
We only copied the HM function and add an if.
We also change the harvest model to dic_harvest and change it to 3 possible values
Then the general function does what it has to do
depending on which we have a model, or a control

The execute model was changed accordingly (the dic thing).

The figure lattice has new and improved figures that account for both models.

To represent the "control we duplicated and add it as a model, but withput harvesting.
This helps to represent it in each of the facets