SO000
this comit is after talk with kevin.

The thing that whas change was adding a random state PER replicate, both in the intial lattice
and in the initla pisiotn of workers
This only conceptual change is that in bth odel, workers, satr in a random plant, independe of the fruit load

So, this will  help to calculate differences ONLY due to the harvestig mode and maybe anova is no t necessary

I have to update funcionts in simple codes. initial conditions was set to replicate only some examples.
And execute model was only change to inclue 2 replicates